Add textfield to category word and small refactoring
