+++ front/src/components/Category.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
+import { useSelector } from 'react-redux'
 
-//Reducers
-import { unAssignWord } from '../reducers/assignReducer'
+//Components
+import CategoryItem from './CategoryItem'
 
 //Mui components
 import Box from '@mui/material/Box'
@@ -10,11 +10,6 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import Divider from '@mui/material/Divider'
 import List from '@mui/material/List'
-import ListItem from '@mui/material/ListItem'
-import ListItemText from '@mui/material/ListItemText'
-import IconButton from '@mui/material/IconButton'
-import DeleteIcon from '@mui/icons-material/Delete'
-
 
 /**
  * Component for category.
@@ -23,16 +18,6 @@ import DeleteIcon from '@mui/icons-material/Delete'
  */
 const Category = ({ category }) => {
   const assignedWords = useSelector(state => state.assignedWords.filter(word => word.category === category))
-  const dispatch = useDispatch()
-
-  /**
-   * Removes the word from the category.
-   * @see unAssignWord
-   * @param {*} word
-   */
-  const handleRemoveFromCategory = (word) => {
-    dispatch(unAssignWord(word))
-  }
 
   return(
     <Grid item>
@@ -56,19 +41,7 @@ const Category = ({ category }) => {
         >
           <List dense>
             {assignedWords.map(word => (
-              <ListItem
-                key={word.word + '-assignItem'}
-                divider
-                secondaryAction={
-                  <IconButton edge='end' onClick={() => handleRemoveFromCategory(word.word)}>
-                    <DeleteIcon/>
-                  </IconButton>
-                }
-              >
-                <ListItemText key={word.word + '-assignText'}>
-                  {word.word}
-                </ListItemText>
-              </ListItem>
+              <CategoryItem key={word.word} word={word.word}/>
             ))}
           </List>
         </Box>
